Update CPT-Archive to work with Timber's get Menu (#14)
Changed the filter that `nav_menu_classes` gets executed on to `wp_get_nav_menu_items` this means that the Classes are added much earlier in the execution.

Timber does not get the menu object at any point, instead going straight for the menu items and assembles the menu itself.  Therefore adding the needed CSS classes to the items, not the objects works in the case.